Add one more generic parameter to AbstractListingResource for the IdType, plus a converter to convert between String and IdType
