+++ .gitignore
@@ -26,3 +26,5 @@ hs_err_pid*
 .project
 .classpath
 /rest-jersey-utils.iml
+
+.idea
